+++ src/components/page/mine/management/ManagementForHost/ManagementHeaderForHost.tsx
@@ -1,5 +1,5 @@
 import { components } from '@/__generated__/schema2';
-import { useDownloadMeetingMemberCSVMutation } from '@api/meeting/hook';
+import { useDownloadMeetingMemberCSVMutation } from '@api/meeting/mutation';
 import DownloadIcon from '@assets/svg/download.svg';
 import { styled } from 'stitches.config';
